+++ examples/shaders.rs
@@ -1,11 +1,13 @@
 use std::borrow::Cow;
 
+use cushy::animation::ZeroToOne;
 use cushy::figures::units::Px;
 use cushy::figures::Rect;
+use cushy::graphics::SimpleRenderOperation;
 use cushy::kludgine::{wgpu, RenderingGraphics};
 use cushy::widget::MakeWidget;
 use cushy::widgets::Canvas;
-use cushy::{Run, SimpleRenderOperation};
+use cushy::Run;
 
 static TRIANGLE_SHADER: &str = r#"
     @vertex
@@ -70,7 +72,12 @@ impl SimpleRenderOperation for TriangleShader {
         Self { pipeline }
     }
 
-    fn render(&self, region: Rect<Px>, _opacity: f32, graphics: &mut RenderingGraphics<'_, '_>) {
+    fn render(
+        &self,
+        region: Rect<Px>,
+        _opacity: ZeroToOne,
+        graphics: &mut RenderingGraphics<'_, '_>,
+    ) {
         println!("Render to {region:?} clipped to {:?}", graphics.clip_rect());
         graphics.pass_mut().set_pipeline(&self.pipeline);
         graphics.pass_mut().draw(0..3, 0..1);